switch from c float to double

+++ src/ch_xs_coords.h
@@ -5,8 +5,8 @@
 
 typedef struct
 {
-  float station;
-  float elevation;
+  double station;
+  double elevation;
 } ChXSCoordinate;
 
 struct ChXSCoords
